Add action to update civicrm group verifications
