+++ decidim-civicrm/config/locales/ca.yml
@@ -21,3 +21,8 @@ ca:
                 is_reserved: Reservat
                 "no": "No"
                 "yes": "Sí"
+                update: Actualitza
+                error: No s'han pogut obtenir els grups per un error a l'API.
+              update:
+                success: S'han actualitzat les verificacions per les usuàries del grup %{group}
+                error: No s'han pogut actualitzat les verificacions per les usuàries del grup %{group}
